requirements.txt
pandas
beautifulsoup4
matplotlib
requests
prisma
geopy
python-dotenv
